pyproject.toml
[tool.poetry]
name = "name-analysis"
version = "0.1.0"
description = ""
authors = ["François Weber <francois.weber.commeercial@gmx.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
matplotlib = "^3.9.0"
unidecode = "^1.3.8"
typer = "^0.12.3"
watchdog = "^4.0.1"
pandas = "^2.2.2"
streamlit-pydantic = {path = "/Users/fweber/code/streamlit-pydantic", develop=true}


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
